+++ js/main.js
@@ -1,5 +1,5 @@
 
-var version="1.2.0";
+var version="1.3.0";
 
 
 var count=0;