+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
  
+## [1.3.0] - 2020-06-12
+- New response type: ``select``
+- Window shows experiment title when loading
+
 ## [1.2.0] - 2020-06-11
 - New response type: ``text``
 - Bug fixes when responding